api/requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
kafka-python==2.0.2
